sonar-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
perplexityai==1.0.0
